processor: abstracting common class functionalities
renaming module names
